Narrow the label vocabulary to u32; give MatchMethod its polarity
Roadmap v0.4.0 item 15. Four of the five review findings it lists (A2,
D1, D2, D3) had already shipped in 78a2c99 with v0.3.0 itself; what was
genuinely open was P5 plus the MatchMethod polarity gap item 12a parked
here.

P5, wider than "buffers":

- The engine's provisional-label, compaction and union-find buffers are
  u32, halving the width of the working set that both full-image
  labeling passes read and write. No wall-clock claim; this is a
  buffer-width result.
- The public label-index vocabulary narrows with them, because the u64
  surface promised indexes no pass could produce: LabelPixel::MAX_LABEL
  / from_label_index / to_label_index, Labeling::label_count, the count
  connected_components_into returns,
  Error::LabelOverflow::label_capacity, and
  ContourHierarchy::component_for_label. Breaking against the published
  v0.3.0, accepted deliberately while the user base is small.
- Two mechanics moved with the width: pass 1's "no neighbour yet"
  sentinel is 0 rather than the integer maximum, because u32::MAX is
  itself a valid label, and exhaustion of the label space is a checked
  LabelOverflow from make_set rather than a "> u32::MAX" comparison a
  u32 can never satisfy.

The polarity hand-off:

- MatchMethod gains the non-generic supertrait ScorePolarity, whose
  EXTREMUM const states where a method's score map marks the best
  match: SAD and SSD say Minimum, NCC says Maximum, and the
  template_match worked example now writes
  interpolate_peak(&scores, best, SSD::EXTREMUM).
- A supertrait rather than the recommended associated const on
  MatchMethod itself, because a const on a three-parameter generic
  trait cannot be read without naming all three parameters. Breaking
  for external implementors only; callers are unaffected.

+++ src/analyze/components/mod.rs
@@ -92,7 +92,7 @@ pub struct Labeling<L: LabelPixel> {
     pub labels: Image<L>,
 
     /// Number of distinct foreground components in `labels`.
-    pub label_count: u64,
+    pub label_count: u32,
 }
 
 impl<L: LabelPixel> core::fmt::Debug for Labeling<L> {
